subindo projeto com metodos deletebycpf e findbycpf, cadastrar embalagem atualizados

+++ src/main/java/com/miniprojeto/miniprojeto/Service/UsuarioService.java
@@ -3,6 +3,7 @@ package com.miniprojeto.miniprojeto.Service;
 import com.miniprojeto.miniprojeto.Model.UsuarioModel;
 import com.miniprojeto.miniprojeto.Repository.UsuarioRepository;
 import com.miniprojeto.miniprojeto.dto.UsuarioDto;
+import com.miniprojeto.miniprojeto.exception.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -44,7 +45,27 @@ public class UsuarioService {
     }
 
     public void deletaUsuario(Long id){
-
+Optional<UsuarioModel> optionalUsuarioModel = usuarioRepository.findById(id);
+if (optionalUsuarioModel.isEmpty()) {
+    throw new ObjectNotFoundException("não foi encontrado um usuário com este registro em nossa base de dados");
+}
         usuarioRepository.deleteById(id);
+    }
+
+    public void deleteByCpf(String cpf) {
+        Optional<UsuarioModel> optionalUsuarioModel = usuarioRepository.findByCpf(cpf);
+        if (optionalUsuarioModel.isEmpty()) {
+            throw new ObjectNotFoundException("este cpf não está registrado");
+        }
+         usuarioRepository.deleteByCpf(cpf);
+    }
+
+    public UsuarioModel findByCpf(String cpf) {
+        Optional<UsuarioModel> optionalUsuarioModel = usuarioRepository.findByCpf(cpf);
+if (optionalUsuarioModel.isEmpty()) {
+    throw new ObjectNotFoundException("não existe");
+}
+
+            return optionalUsuarioModel.get();
     }
 }
